requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pika==1.3.2
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7
pydantic==2.5.1
asyncio==3.4.3
websockets==12.0
python-dotenv==1.0.0